Langgraph/requirements.txt
ipykernel

langgraph
notebook
python-dotenv

langchain
langchain-google-genai
pydantic